refactor: customize 파일명 변경, request 타입 추가
