+++ src/apis/requests/debateTable.ts
@@ -1,0 +1,5 @@
+import { DebateTableData } from '../../type/type';
+
+export interface PutDebateTableRequestType extends DebateTableData {
+  id: number;
+}
