+++ src/apis/responses/debateTable.ts
