+++ src/hooks/query/useGetDebateTableData.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { getRepository } from '../../repositories/DebateTableRepository';
-import { GetDebateTableResponseType } from '../../apis/responses/customize';
+import { GetDebateTableResponseType } from '../../apis/responses/debateTable';
 
 export function useGetDebateTableData(tableId: number, enabled?: boolean) {
   return useQuery<GetDebateTableResponseType>({
